Create guessing.c
In this program, the computer will randomly select any number and the user has to guess the number. In case of wrong guesses the computer will give hints (whether number is lesser or greater than the guessed number). In the end of execution , when the user have guessed the correct number then the computer will count and tell the number of guesses taken by the user to guess the correct number.